move block context handling to its own module
